LayoutTests/http/wpt/webauthn/public-key-credential-get-failure-nfc.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' within user activated events.

PASS PublicKeyCredential's [[get]] with timeout in a mock nfc authenticator.
